pyproject.toml
[project]
name = "quest"
version = "0.1.0"
description = "Question Engine for Quiz Games"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "flask>=3.1",
    "flask-sqlalchemy>=3.1",
    "flask-cors>=5.0",
    "pydantic>=2.0",
    "pyjwt>=2.9",
    "flasgger>=0.9",
    "gunicorn>=23.0",
    "requests>=2.31",
    "strands-agents>=1.0",
    "strands-agents-tools>=0.2",
    "bedrock-agentcore>=1.0",
    "boto3>=1.35",
    "python-dotenv>=1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
